// src/app/commonBussinessSRC/psComponents/ps-lookup/ps-lookup-beneficiaries-local/ps-lookup-beneficiaries-local.component.interfaces.ts
import { Type } from '@angular/core';
import { IOptionsPsBaseFieldExposed } from 'src/app/commonSRC/psServices/models/ps-common-interface';

export interface IOptionsPsLookupBenificiariesLocalExposed extends IOptionsPsBaseFieldExposed {
    listOfBenificiaries?: any[];
    currency?: string;
    component?: Type<any>;
    showOnlyList?: boolean;
    type?: string;
    requestObject: any;
}